+++ src/QueryDescription/index.d.ts
@@ -16,6 +16,7 @@ declare module '@nozbe/watermelondb/QueryDescription' {
     | 'oneOf'
     | 'notIn'
     | 'between'
+    | 'match'
 
   export interface ColumnDescription {
     column: ColumnName
@@ -103,6 +104,9 @@ declare module '@nozbe/watermelondb/QueryDescription' {
   export function between(left: number, right: number): Comparison
   export function column(name: ColumnName): ColumnDescription
   export function where(left: ColumnName, valueOrComparison: Value | Comparison): WhereDescription
+  export function and(...conditions: Where[]): And
+  export function or(...conditions: Where[]): Or
+  export function textMatches(value: string): Comparison
   export function and(...conditions: Condition[]): And
   export function or(...conditions: Condition[]): Or
   export function like(value: string): Comparison