+++ CHANGELOG.md
@@ -196,6 +196,9 @@ cause all or some of user's data to be deleted. More information available via G
   This is called when underlying IndexedDB encountered a quota exceeded error (ran out of allotted disk space for app)
   This means that app can't save more data or that it will fall back to using in-memory database only
   Note that this only works when `useWebWorker: false`
+- [SQLiteAdapter] Added support for Full Text Search for SQLite adapter:
+  Add `isSearchable` boolean flag to schema column descriptor for creating Full Text Search-able columns
+  Add `Q.textMatches(value)` that compiles to `match 'value'` SQL for performing Full Text Search using SQLite adpater
 
 ### Changes